Move GUS class and constant definitions to header file
